Split common signals off app.py class so that modules can connect to them
In order for modules (none of which is currently implemented) to monitor
for changes in the main interface some signals need to be split off from
app.py into a centralised class. Module `signals` will now contain all
common signals modules are supposed to observe or emit that are not
specific to the main interface itself.