Add support for experiment results with more than one data tables
HDF5 datastore now supports experiment results with more than one data
tables. Instead of having a singular dataset per experiment a group
can be created instead which can contain more than one dataset in any
arrangement. This is done via the `make_wb_group` and
`make_synthetic_group` functions which mirror the semantics of the
corresponding single dataset files. The group or the group name can then
be passed to the new `grp` argument in `make_wb_table` and
`make_synthetic_table` to nest the results under the group (which must
preexist). This does not affect the handling of the data from arc2control
modules. If a module is producing a group instead of a single table then
the group itself is passed as an argument when instantiating the display
widget (if any).